Rework condition action and add two more conditions
- Reworked entire setup to be more compact and less convoluted
- Added vJoy and logical device as condition options
