--- action_plugins/condition/DirectionComparatorUI.qml
@@ -1,47 +1,0 @@
-// -*- coding: utf-8; -*-
-//
-// Copyright (C) 2015 - 2022 Lionel Ott
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-
-import QtQuick
-import QtQuick.Controls
-import QtQuick.Layouts
-
-import Gremlin.ActionPlugins
-
-import "../../qml"
-
-
-Item {
-    id: _root
-
-    property DirectionComparator comparator
-
-    implicitHeight: _content.height
-    implicitWidth: _content.width
-
-    ColumnLayout {
-        id: _content
-
-        Label {
-            text: "Any of the followng directions is pressed"
-        }
-
-        HatDirectionSelectorV2 {
-            directions: comparator.model
-        }
-    }
-}
