+++ qml/BootstrapIconsNames.qml
@@ -39,6 +39,7 @@ QtObject {
     readonly property string script: "\uF2C6"
     readonly property string instance: "\uF227"
     readonly property string configure: "\uF56B"
+    readonly property string bullet_point: "\uF309"
 
     readonly property string icon_mouse: "\uF49D"
     readonly property string icon_keyboard: "\uF451"
